{{cookiecutter.project_name}}/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-redis
django-cors-headers
django-environ
django-ckeditor
Pillow
daphne